[Rust Frontend] Validate preprocessed multimodal features in chat
Replace InlineMmFeatures with MmFeatures and validate features at the prepare_preprocessed entry point. Move MessagePack decoding and payload limits into the gRPC adapter, and relocate validation tests to their owning layers.

Validation: 297 relevant nextest cases, Clippy with warnings denied, and rustfmt passed.
