+++ rust/src/chat/src/error.rs
@@ -27,8 +27,8 @@ pub enum Error {
     UnsupportedModality { modality: String },
     #[error("At most {limit} {modality}(s) may be provided in one prompt.")]
     MmLimitExceeded { modality: String, limit: usize },
-    #[error(transparent)]
-    InlineMultimodal(#[from] vllm_engine_core_client::protocol::multimodal::InlineMmError),
+    #[error("invalid inline multimodal features: {message}")]
+    InvalidPreprocessedMultimodal { message: String },
     #[error("multimodal preprocessing error: {0}")]
     Multimodal(#[message] String),
     #[error("{kind} parsing is not available for model `{model_id}`")]
@@ -108,7 +108,7 @@ impl Error {
             Self::UnsupportedMultimodalRenderer
             | Self::UnsupportedMultimodalContent(_)
             | Self::UnsupportedModality { .. }
-            | Self::InlineMultimodal(_)
+            | Self::InvalidPreprocessedMultimodal { .. }
             | Self::MmLimitExceeded { .. } => true,
 
             _ => false,
